// sanity.types.ts
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: ./sanity/queries/query.ts
// Variable: BRANDS_QUERY
// Query: *[_type=='brand'] | order(name asc)
export type BRANDS_QUERYResult = Array<{
  _id: string;
  _type: "brand";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  description?: string;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
}>;